chore: clippy fixes

+++ src/semantic_analysis/type_registry.rs
@@ -87,9 +87,9 @@ impl TypeRegistry {
                 .map(|t| Type::MutPointer(Box::new(t))),
             grammar::Type::Array(t, size) => self
                 .resolve_grammar_type(scope, t.as_ref())
-                .map(|t| Type::Array(Box::new(t), *size as usize)),
+                .map(|t| Type::Array(Box::new(t), *size)),
             grammar::Type::Ident(ident) => self.resolve_string(scope, ident.as_str()),
-            grammar::Type::Unknown(size) => Some(self.padding_type(*size as usize)),
+            grammar::Type::Unknown(size) => Some(self.padding_type(*size)),
         }
     }
